darray: change element_size to size_t
This argument is usually the result of a sizeof().

+++ libnopegl/src/darray.c
@@ -56,7 +56,7 @@ static int reserve_aligned(struct darray *darray, int capacity)
     return 0;
 }
 
-void ngli_darray_init(struct darray *darray, int element_size, int aligned)
+void ngli_darray_init(struct darray *darray, size_t element_size, int aligned)
 {
     darray->data = NULL;
     darray->count = 0;
